\name{CGHbase-package}
\alias{CGHbase-package}
\alias{CGHbase}
\docType{package}
\title{CGHbase: Base functions and classes for arrayCGH data analysis.}
\description{CGHbase: Base functions and classes for arrayCGH data analysis.}
\details{

	Main infrastructural classes: \code{\link{cghRaw}}, \code{\link{cghSeg}}, \code{\link{cghCall}}.
	Full help on methods and associated functions is available from withing class help pages.
	
	Attached data sets: \code{\link{WiltingData}}, \code{\link{WiltingRaw}}, \code{\link{WiltingNorm}}, \code{\link{WiltingSeg}}, \code{\link{WiltingCalled}}.
	
}
\author{Sjoerd Vosse <sjoerdvos@yahoo.com>}
\keyword{package}

