\name{plot.cghRaw}
\alias{plot.cghRaw}
\alias{plot.cghSeg}
\alias{plot.cghCall}
\alias{plot.cghRegions}

\title{Plot aCGH data.}

\description{
  Please see the class descriptions for more details on the plot functions.
}

\usage{
  \method{plot}{cghRaw}(x, y, \dots)
  \method{plot}{cghSeg}(x, y, \dots)
  \method{plot}{cghCall}(x, y, \dots)
  \method{plot}{cghRegions}(x, y, \dots)
}

\arguments{
  \item{x}{ An object of class \code{\link{cghRaw}}, \code{\link{cghSeg}}, \code{\link{cghCall}}, or \code{\link{cghSeg}}.}
  \item{y}{ This argument is not used and should be missing.}
  \item{\dots}{ Arguments \code{plot}. }
}

\author{Sjoerd Vosse}

\seealso{\code{\link{cghRaw-class}}, \code{\link{cghSeg-class}}, \code{\link{cghCall-class}}, \code{\link{cghRegions-class}}}

