\name{WiltingCalled}
\docType{data}
\alias{WiltingCalled}
\title{Cervical cancer arrayCGH data called with CGHcall}
\description{
  Cervical cancer arrayCGH data called with \code{\link[CGHcall]{CGHcall}} with default settings, containing 3552 features for 5 samples.
}
\usage{WiltingCalled}
\format{
  An object of class \code{\link{cghCall}}
}
\source{
  Wilting, S.M., Snijders, P.J., Meijer, G.A., Ylstra, B., van den IJssel, P.R., Snijders, A.M., Albertson, D.G., Coffa, J., Schouten, J.P., van de Wiel, M.A., Meijer, C.J., & Steenbergen, R.D. (2006). Increased gene copy numbers at chromosome 20q are frequent in both squamous cell carcinomas and adenocarcinomas of the cervix. \emph{Journal of Pathology, 210}, 258-259.
  
  Mark A. van de Wiel, Kyung In Kim, Sjoerd J. Vosse, Wessel N. van Wieringen, Saskia M. Wilting and Bauke Ylstra. CGHcall: calling aberrations for array CGH tumor profiles. \emph{Bioinformatics, 23}, 892-894.
}
\keyword{datasets}