\name{WiltingRaw}
\docType{data}
\alias{WiltingRaw}
\title{Raw log2 ratios from cervical cancer arrayCGH data.}
\description{
  Raw log2 ratios from cervical cancer arrayCGH data, containing 3552 features for 5 samples. These data have been preprocessed using \code{\link[CGHcall]{preprocess}}.
}
\usage{WiltingCalled}
\format{
  An object of class \code{\link{cghRaw}}.
}
\source{
  Wilting, S.M., Snijders, P.J., Meijer, G.A., Ylstra, B., van den IJssel, P.R., Snijders, A.M., Albertson, D.G., Coffa, J., Schouten, J.P., van de Wiel, M.A., Meijer, C.J., & Steenbergen, R.D. (2006). Increased gene copy numbers at chromosome 20q are frequent in both squamous cell carcinomas and adenocarcinomas of the cervix. \emph{Journal of Pathology, 210}, 258-259.
}
\keyword{datasets}