\name{WiltingRegions}
\docType{data}
\alias{WiltingRegions}
\title{Regions of cervical cancer arrayCGH data as defined by CGHregions}
\description{
  Regions of cervical cancer arrayCGH data as defined by \code{\link[CGHregions]{CGHregions}} with default settings, containing 90 regions over 5 samples.
}
\usage{WiltingRegions}
\format{
  An object of class \code{\link{cghRegions}}
}
\source{
  Wilting, S.M., Snijders, P.J., Meijer, G.A., Ylstra, B., van den IJssel, P.R., Snijders, A.M., Albertson, D.G., Coffa, J., Schouten, J.P., van de Wiel, M.A., Meijer, C.J., & Steenbergen, R.D. (2006). Increased gene copy numbers at chromosome 20q are frequent in both squamous cell carcinomas and adenocarcinomas of the cervix. \emph{Journal of Pathology, 210}, 258-259.
  
  Mark A. van de Wiel and Wessel N. van Wieringen (2007). CGHregions: Dimension Reduction for Array CGH Data with Minimal Information Loss. \emph{Cancer Informatics, 2}, 55-63.
}
\keyword{datasets}